fix: refactor NavUser and LoginSidebarMenu components feat: implement NavUserDropdown for enhanced user profile interactions fix: update getProfileFormData to conditionally fetch identities data fix: enhance NewFeatureBadge hover effect
